+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [0.0.3]
+
+- Differentiate between zero results and failing to run any greps.
+- Reveal and focus the results panel when a search is complete.
+- Thanks to [Anton Kudris](https://github.com/jodaka) for reporting these issues.
+
 ## [0.0.2]
 
 - Icon update.